Converted orphan buffer to use proxy wiring. Not yet completed.
