+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/builder/PlatformComponentBuilder.java
@@ -26,6 +26,8 @@ import com.swirlds.platform.event.deduplication.EventDeduplicator;
 import com.swirlds.platform.event.deduplication.StandardEventDeduplicator;
 import com.swirlds.platform.event.hashing.DefaultEventHasher;
 import com.swirlds.platform.event.hashing.EventHasher;
+import com.swirlds.platform.event.orphan.DefaultOrphanBuffer;
+import com.swirlds.platform.event.orphan.OrphanBuffer;
 import com.swirlds.platform.event.validation.DefaultEventSignatureValidator;
 import com.swirlds.platform.event.validation.DefaultInternalEventValidator;
 import com.swirlds.platform.event.validation.EventSignatureValidator;
@@ -64,6 +66,7 @@ public class PlatformComponentBuilder {
     private EventDeduplicator eventDeduplicator;
     private EventSignatureValidator eventSignatureValidator;
     private StateGarbageCollector stateGarbageCollector;
+    private OrphanBuffer orphanBuffer;
 
     /**
      * False if this builder has not yet been used to build a platform (or platform component builder), true if it has.
@@ -296,5 +299,37 @@ public class PlatformComponentBuilder {
             stateGarbageCollector = new DefaultStateGarbageCollector(blocks.platformContext());
         }
         return stateGarbageCollector;
+    }
+
+    /**
+     * Provide an orphan buffer in place of the platform's default orphan buffer.
+     *
+     * @param orphanBuffer the orphan buffer to use
+     * @return this builder
+     */
+    @NonNull
+    public PlatformComponentBuilder withOrphanBuffer(@NonNull final OrphanBuffer orphanBuffer) {
+        throwIfAlreadyUsed();
+        if (this.orphanBuffer != null) {
+            throw new IllegalStateException("Orphan buffer has already been set");
+        }
+        this.orphanBuffer = Objects.requireNonNull(orphanBuffer);
+        return this;
+    }
+
+    /**
+     * Build the orphan buffer if it has not yet been built. If one has been provided via
+     * {@link #withOrphanBuffer(OrphanBuffer)}, that orphan buffer will be used. If this method is called more than
+     * once, only the first call will build the orphan buffer. Otherwise, the default orphan buffer will be created and
+     * returned.
+     *
+     * @return the orphan buffer
+     */
+    @NonNull
+    public OrphanBuffer buildOrphanBuffer() {
+        if (orphanBuffer == null) {
+            orphanBuffer = new DefaultOrphanBuffer(blocks.platformContext(), blocks.intakeEventCounter());
+        }
+        return orphanBuffer;
     }
 }
